Port code to use boost::python::numpy
Support for boost::python::numeric has been dropped in boost v1.65, code
must now directly tap into boost::python::numpy. The API is quite
different, so this has a significant impact VPython's code.
Mostly it involves switching from boost::python::numeric::array to
boost::python::numpy::ndarray, and adapting code to use ndarray's API.
Also VPython must now build against the boost_numpy library too.

The current state of affairs is that the code compiles, but segfaults
upon `import visual` in a Python interpreter. This may or may not be due
to wxPython, or something else. I'm on a grsec kernel for the time being
and can't debug easily to determine the cause of the crash.

I should note that some of the changes made to the code were not due
solely to API changes in Boost:
* added `Py_Initialize()` in `init_numpy()`
* use newer versions of obsolete Numpy types/functions/constants:
 * PyArray_DOUBLE -> NPY_DOUBLE
 * PyArray_FromDims() -> PyArray->SimpleNew()
 * ((PyArrayObject*)ptr)->data -> PyArray_DATA(ptr)
 * NPY_ENSUREARRAY -> NPY_ARRAY_ENSUREARRAY
 * NPY_CONTIGUOUS -> NPY_ARRAY_C_CONTIGUOUS
 * NPY_ALIGNED -> NPY_ARRAY_ALIGNED